PM: DDS Updates PGPE (2/2)
HW-Image-Prereq: I0e1ae888f795226f3eed59d92fd97ed07982e9b5
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/107743
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Dev-Ready: RAHUL BATRA <rbatra@us.ibm.com>

+++ import/chips/p10/procedures/ppe/pgpe/pgpe_gppb.h
@@ -105,9 +105,9 @@ uint8_t pgpe_gppb_get_dds_trip_ps_slope(uint32_t pt_set, uint32_t core, uint32_t
 uint32_t pgpe_gppb_get_dds_delay(uint32_t core, uint32_t idx);
 uint32_t pgpe_gppb_get_dds_trip(uint32_t core, uint32_t idx);
 uint32_t pgpe_gppb_get_dds_cal_adj(uint32_t core, uint32_t idx);
-uint32_t pgpe_gppb_get_dds_other_droop_count_control();
-uint32_t pgpe_gppb_get_dds_other_ftc_large_droop_mode_reg_setting();
-uint32_t pgpe_gppb_get_dds_other_ftc_misc_droop_mode_reg_setting();
+uint64_t pgpe_gppb_get_dds_other_droop_count_control();
+uint64_t pgpe_gppb_get_dds_other_ftc_large_droop_mode_reg_setting();
+uint64_t pgpe_gppb_get_dds_other_ftc_misc_droop_mode_reg_setting();
 uint32_t pgpe_gppb_get_dds_trip_mode();
 uint32_t pgpe_gppb_get_dds_trip_intp_ctrl(uint32_t x);
